+++ src/components/DeviceFormDialog.jsx
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 import { useState, useEffect, useRef } from "react";
-import { STYLES } from "../service/global_variable";
+import { STYLES } from "../services/global_variable";
 import {
   Dialog,
   DialogTitle,
