+++ src/services/global_function.jsx
