+++ src/services/global_variable.jsx
