+++ src/services/widget_service.js
